requirements.txt
Flask==2.1.1
h5py==3.6.0
numpy==1.21.6
Pillow==9.1.0
tensorflow==2.8.0
python-dotenv==0.19.2
gunicorn==20.1.0
gunicorn[gevent]==20.1.0
